requirements.txt
kafka-python
pyspark==2.4.7
requests